Rename and move combined service
