+++ src/vs/workbench/contrib/customEditor/common/contributedCustomEditors.ts
@@ -11,8 +11,8 @@ import { IExtensionDescription } from 'vs/platform/extensions/common/extensions'
 import { IStorageService, StorageScope, StorageTarget } from 'vs/platform/storage/common/storage';
 import { Memento } from 'vs/workbench/common/memento';
 import { CustomEditorDescriptor, CustomEditorInfo } from 'vs/workbench/contrib/customEditor/common/customEditor';
-import { ContributedEditorPriority } from 'vs/workbench/contrib/customEditor/common/extensionContributedEditorService';
 import { customEditorsExtensionPoint, ICustomEditorsExtensionPoint } from 'vs/workbench/contrib/customEditor/common/extensionPoint';
+import { ContributedEditorPriority } from 'vs/workbench/services/editor/common/editorOverrideService';
 import { IExtensionPointUser } from 'vs/workbench/services/extensions/common/extensionsRegistry';
 
 export const defaultCustomEditor = new CustomEditorInfo({
